finished detailed element view UI

+++ Liqbo/Models/ProductDataModel.swift
@@ -22,7 +22,7 @@ class ProductDataModel{
     var package: String = ""
     var total_package_units: Int = 0
     var volume_in_milliliters: Int = 0
-    var alcohol_content: Int = 0
+    var alcohol_content: Float = 0
     var style = ""
     var image_thumb_url : URL = URL(string: "https://s3.amazonaws.com/woof.nextglass.co/custom_item_type_images_production/af7e5a9d2a54837d65cae5637975ba8dfa05311f-custom-item-type-image.png?1476365129")!
     var image_url : URL = URL(string: "https://s3.amazonaws.com/woof.nextglass.co/custom_item_type_images_production/af7e5a9d2a54837d65cae5637975ba8dfa05311f-custom-item-type-image.png?1476365129")!
